Codebase simplification
