+++ .github/workflows/smoketests.yml
@@ -7,6 +7,10 @@ on:
 jobs:
   tests:
     runs-on: ubuntu-20.04
+    services:
+      redis:
+        image: redis
+        container_name: beasty_redis
     steps:
     - uses: actions/checkout@v2
     - uses: actions/setup-go@v2
