--- README.md
@@ -1,26 +1,0 @@
-# discord-beasty-bot
---
-
-## Summary
---
-
-## Layout
-TBD (Pending changes)
-
-## Environment Setup
---
-
-## Integrations
---
-
-### Discord
---
-
-### Sendgrid
---
-
-### CodeClimate
---
-
-## Services
---
