+++ cmd/beasty/responses/responses.go
@@ -1,4 +1,4 @@
-package main
+package responses
 
 import (
 	"log"
@@ -6,21 +6,21 @@ import (
 	"time"
 )
 
-type responses struct {
+type Responses struct {
 	modelData map[interface{}]interface{}
 }
 
 // responses fetches a random string based on a lookup
-func newResponses(model map[interface{}]interface{}) *responses {
+func NewResponses(model map[interface{}]interface{}) *Responses {
 	rand.Seed(time.Now().UnixNano())
-	r := &responses{
+	r := &Responses{
 		modelData: model,
 	}
 	return r
 }
 
 // returns response string randomly selected based on lookup
-func (r responses) generateResponse(lookup string) string {
+func (r Responses) GenerateResponse(lookup string) string {
 	// fetch map and cast to array of strings
 	respMap, valid := r.modelData[lookup].([]interface{})
 	if !valid {
