+++ cmd/beasty/responses/responses.yaml
@@ -7,19 +7,17 @@ startup:
 shutdown:
   - "Brb gotta sleep."
   - "I don't feel so good Mr. Stark."
+  - "Bedtime"
 welcome:
   - "Welcome to the club! We have code and... lots of code!"
   - "Welcome!"
   - "Welcome to the party!"
   - "Howdy!"
 student:
-  - "I've sent you an email through your @viu.ca virtual email, send me the code via the code command"
-email_invalid:
-  - "Hey it looks like I couldn't parse that as a valid viu email. Sorry about that!"
-student_role:
-  - "Hey! thanks for submitting a valid code, you now have the student role."
-code_invalid:
-  - "Heya, looks like I don't have that as a valid code in my data store! Could you generate one via the student command?"
+  - "You now have the student role!"
+  - "Poof! Behold, the student role."
+  - "Ok you're a student now but can you write perl?"
+  - "Hey you're a student now. Don't forget to drop all the tables!"
 joke:
   - "Where is John Connor?"
   - "Try harder Phil."
@@ -32,7 +30,7 @@ joke:
   - "There are no runtime errors, just happy little segmentation faults"
   - "(sorry (dont (mind (my (lisp)))));"
 unkown:
-  - "Sorry I don't understand that command, try the help command"
+  - "Sorry I don't understand that command, try the 'help' command"
 help:
-  - "Recognized commands: help, joke, student, code"
+  - "Recognized commands: help, joke, student"
 ...
