--- config.yml
@@ -1,17 +1,0 @@
----
-# You can get the IDs of channels and add them to this list so the bot knows about them
-# TODO: May need to consider data storage for this list as channels will become more dynamic for classes
-watch_channels:
-  - "747583190813900850"
-  - "999999999999999999"
-watch_dms: true
-startup_channel_id: "747583190813900850" # where the bot displays its startup message
-join_channel_id: "747583190813900850" # channel to watch new users joining
-student_role_id: "752361111789109343"
-guild_id: "662891783981891584"
-signal_word: "/beasty"
-responses_file: "responses.yml" # Response model data location
-sender_email: "noreply@viu-csci.pye.dev"
-email_subject: "VIU-CSCI Student Role Verification Code"
-sender_message: "Hey %s! Here's a verification code for you to send me on discord: %d"
-...
