+++ go.mod
@@ -1,15 +1,9 @@
 module github.com/viu-csci-guild/beasty
 
-require (
-	github.com/bwmarrin/discordgo v0.22.0
-	github.com/go-redis/redis/v8 v8.0.0-beta.8 // indirect
-	github.com/go-yaml/yaml v2.1.0+incompatible
-	github.com/gomodule/redigo v1.8.2
-	github.com/securego/gosec/v2 v2.6.1 // indirect
-	github.com/sendgrid/rest v2.6.1+incompatible // indirect
-	github.com/sendgrid/sendgrid-go v3.6.3+incompatible
-	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
-	golang.org/x/tools v0.1.0 // indirect
-)
+go 1.16
 
-go 1.15
+require (
+	github.com/bwmarrin/discordgo v0.23.2
+	github.com/go-yaml/yaml v2.1.0+incompatible
+	github.com/gomodule/redigo v1.8.5
+)
