--- storage.go
@@ -1,10 +1,0 @@
-package main
-
-type storage struct {
-	redis struct{}
-}
-
-func NewStorage() *storage {
-	s := &storage{}
-	return s
-}
